+++ test/lightning_web/live/work_order_live_test.exs
@@ -431,16 +431,22 @@ defmodule LightningWeb.RunWorkOrderTest do
              |> element("input#run-search-form_search_term")
              |> has_element?()
 
-      ## both log and body select
+      ## all id, log and body select
       assert view
              |> element(
-               "input#run-both-search-form_body[type='hidden'][value='true']"
+               "input#run-all-search-form_body[type='hidden'][value='true']"
              )
              |> has_element?()
 
       assert view
              |> element(
-               "input#run-both-search-form_log[type='hidden'][value='true']"
+               "input#run-all-search-form_log[type='hidden'][value='true']"
+             )
+             |> has_element?()
+
+      assert view
+             |> element(
+               "input#run-all-search-form_id[type='hidden'][value='true']"
              )
              |> has_element?()
 
@@ -457,6 +463,12 @@ defmodule LightningWeb.RunWorkOrderTest do
              )
              |> has_element?()
 
+      assert view
+             |> element(
+               "input#run-body-search-form_id[type='hidden'][value='false']"
+             )
+             |> has_element?()
+
       ## individual search for log
       assert view
              |> element(
@@ -466,7 +478,32 @@ defmodule LightningWeb.RunWorkOrderTest do
 
       assert view
              |> element(
+               "input#run-log-search-form_id[type='hidden'][value='false']"
+             )
+             |> has_element?()
+
+      assert view
+             |> element(
                "input#run-log-search-form_log[type='hidden'][value='true']"
+             )
+             |> has_element?()
+
+      ## individual search for id
+      assert view
+             |> element(
+               "input#run-id-search-form_body[type='hidden'][value='false']"
+             )
+             |> has_element?()
+
+      assert view
+             |> element(
+               "input#run-id-search-form_log[type='hidden'][value='false']"
+             )
+             |> has_element?()
+
+      assert view
+             |> element(
+               "input#run-id-search-form_id[type='hidden'][value='true']"
              )
              |> has_element?()
     end